+++ lib/scales_cms/cms/cms_pages.ex
@@ -23,6 +23,25 @@ defmodule ScalesCms.Cms.CmsPages do
   end
 
   @doc """
+  Returns the list of paginated cms_pages.
+
+  ## Examples
+
+      iex> list_paginated_cms_pages(1, 25)
+      [%CmsPage{}, ...]
+
+  """
+  def list_paginated_cms_pages(page, amount) do
+    offset = page * amount
+
+    CmsPage
+    |> where([cp], is_nil(cp.cms_directory_id))
+    |> limit(^amount)
+    |> offset(^offset)
+    |> repo().all()
+  end
+
+  @doc """
   Returns the list of cms_pages within a directory.
 
   ## Examples
@@ -34,6 +53,21 @@ defmodule ScalesCms.Cms.CmsPages do
   def list_pages_for_directory_id(directory_id) do
     CmsPage
     |> where([cp], cp.cms_directory_id == ^directory_id)
+    |> repo().all()
+  end
+
+  @doc """
+  Returns the list of searched pages.
+
+  ## Examples
+
+      iex> list_pages_for_directory_id("Page title")
+      [%CmsPage{}, ...]
+
+  """
+  def search_pages(query) do
+    CmsPage
+    |> where([cp], ilike(cp.title, ^"%#{query}%"))
     |> repo().all()
   end