+++ lib/scales_cms_web/components/live/cms_components/image_button/image_button_editor.ex
@@ -88,7 +88,12 @@ defmodule ScalesCmsWeb.Components.CmsComponents.ImageButton.ImageButtonEditor do
           <.input type="text" field={@form[:subtitle]} label="Subtitle" />
           <.input type="text" field={@form[:icon]} label="Icon" />
 
-          <.input type="text" field={@form[:page_id]} label="Page ID" />
+          <.live_component
+            id={"page-input-#{@block.id}"}
+            module={ScalesCmsWeb.Components.HelperComponents.PageSearch}
+            field={@form[:page_id]}
+          />
+
           <.input type="text" field={@form[:url]} label="URL" />
           <.input type="textarea" field={@form[:payload]} label="Payload" />